reddit_stream/requirements.txt
praw
loguru
kafka-python
python-dotenv
click
pylint
black